+++ pkg/imageutil/format.go
