+++ pkg/imageutil/shared.go
