+++ pkg/imageutil/stat.go
