+++ pkg/imageutil/tree.go
